refactor: web 모듈로 컨트롤러 이동
