+++ kitchenpos-web/build.gradle
@@ -1,0 +1,4 @@
+dependencies {
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+    testImplementation 'io.rest-assured:rest-assured:3.3.0'
+}
