+++ kitchenpos-web/src/main/java/kitchenpos/menu/MenuGroupRestController.java
@@ -1,0 +1,35 @@
+package kitchenpos.menu;
+
+import java.net.URI;
+import java.util.List;
+import kitchenpos.dto.menu.MenuGroupRequest;
+import kitchenpos.dto.menu.MenuGroupResponse;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class MenuGroupRestController {
+
+    private final MenuGroupService menuGroupService;
+
+    public MenuGroupRestController(final MenuGroupService menuGroupService) {
+        this.menuGroupService = menuGroupService;
+    }
+
+    @PostMapping("/api/menu-groups")
+    public ResponseEntity<MenuGroupResponse> create(@RequestBody final MenuGroupRequest menuGroup) {
+        final MenuGroupResponse created = menuGroupService.create(menuGroup.toEntity());
+        final URI uri = URI.create("/api/menu-groups/" + created.getId());
+        return ResponseEntity.created(uri)
+            .body(created);
+    }
+
+    @GetMapping("/api/menu-groups")
+    public ResponseEntity<List<MenuGroupResponse>> list() {
+        return ResponseEntity.ok()
+            .body(menuGroupService.list());
+    }
+}
