+++ kitchenpos-web/src/main/java/kitchenpos/menu/MenuRestController.java
@@ -1,0 +1,34 @@
+package kitchenpos.menu;
+
+import java.net.URI;
+import java.util.List;
+import kitchenpos.dto.menu.MenuRequest;
+import kitchenpos.dto.menu.MenuResponse;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class MenuRestController {
+    private final MenuService menuService;
+
+    public MenuRestController(final MenuService menuService) {
+        this.menuService = menuService;
+    }
+
+    @PostMapping("/api/menus")
+    public ResponseEntity<MenuResponse> create(@RequestBody final MenuRequest menu) {
+        final MenuResponse created = menuService.create(menu.toEntity());
+        final URI uri = URI.create("/api/menus/" + created.getId());
+        return ResponseEntity.created(uri)
+                .body(created);
+    }
+
+    @GetMapping("/api/menus")
+    public ResponseEntity<List<MenuResponse>> list() {
+        return ResponseEntity.ok()
+                .body(menuService.list());
+    }
+}
