+++ kitchenpos-web/src/main/java/kitchenpos/product/ProductRestController.java
@@ -1,0 +1,34 @@
+package kitchenpos.product;
+
+import java.net.URI;
+import java.util.List;
+import kitchenpos.dto.product.ProductRequest;
+import kitchenpos.dto.product.ProductResponse;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class ProductRestController {
+    private final ProductService productService;
+
+    public ProductRestController(final ProductService productService) {
+        this.productService = productService;
+    }
+
+    @PostMapping("/api/products")
+    public ResponseEntity<ProductResponse> create(@RequestBody final ProductRequest product) {
+        final ProductResponse created = productService.create(product.toEntity());
+        final URI uri = URI.create("/api/products/" + created.getId());
+        return ResponseEntity.created(uri)
+                .body(created);
+    }
+
+    @GetMapping("/api/products")
+    public ResponseEntity<List<ProductResponse>> list() {
+        return ResponseEntity.ok()
+                .body(productService.list());
+    }
+}
