+++ settings.gradle
@@ -1,4 +1,3 @@
 rootProject.name = 'kitchenpos'
 
-include 'kitchenpos-common', 'kitchenpos-application'
-include 'kitchenpos-web'
+include 'kitchenpos-common', 'kitchenpos-application', 'kitchenpos-web'
