--- src/main/java/kitchenpos/product/ui/ProductRestController.java
@@ -1,35 +1,0 @@
-package kitchenpos.product.ui;
-
-import java.net.URI;
-import java.util.List;
-import kitchenpos.product.application.ProductService;
-import kitchenpos.product.dto.ProductRequest;
-import kitchenpos.product.dto.ProductResponse;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class ProductRestController {
-    private final ProductService productService;
-
-    public ProductRestController(final ProductService productService) {
-        this.productService = productService;
-    }
-
-    @PostMapping("/api/products")
-    public ResponseEntity<ProductResponse> create(@RequestBody final ProductRequest product) {
-        final ProductResponse created = productService.create(product.toEntity());
-        final URI uri = URI.create("/api/products/" + created.getId());
-        return ResponseEntity.created(uri)
-                .body(created);
-    }
-
-    @GetMapping("/api/products")
-    public ResponseEntity<List<ProductResponse>> list() {
-        return ResponseEntity.ok()
-                .body(productService.list());
-    }
-}
