--- src/test/java/kitchenpos/utils/DatabaseCleanup.java
@@ -1,51 +1,0 @@
-package kitchenpos.utils;
-
-import com.google.common.base.CaseFormat;
-import java.util.List;
-import java.util.stream.Collectors;
-import javax.persistence.Entity;
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
-import javax.persistence.Table;
-import org.springframework.beans.factory.InitializingBean;
-import org.springframework.stereotype.Service;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@ActiveProfiles("test")
-public class DatabaseCleanup implements InitializingBean {
-
-    @PersistenceContext
-    private EntityManager entityManager;
-
-    private List<String> tableNames;
-
-    @Override
-    public void afterPropertiesSet() {
-        tableNames = entityManager.getMetamodel().getEntities().stream()
-            .filter(e -> e.getJavaType().getAnnotation(Entity.class) != null)
-            .map(e -> {
-                if (e.getJavaType().getAnnotation(Table.class) != null) {
-                    return CaseFormat.UPPER_CAMEL.to(CaseFormat.LOWER_UNDERSCORE,
-                        e.getJavaType().getAnnotation(Table.class).name());
-                }
-                return CaseFormat.UPPER_CAMEL.to(CaseFormat.LOWER_UNDERSCORE, e.getName());
-            })
-            .collect(Collectors.toList());
-    }
-
-    @Transactional
-    public void execute() {
-        entityManager.flush();
-        entityManager.createNativeQuery("SET REFERENTIAL_INTEGRITY FALSE").executeUpdate();
-
-        for (String tableName : tableNames) {
-            entityManager.createNativeQuery("TRUNCATE TABLE " + tableName).executeUpdate();
-            entityManager.createNativeQuery(
-                "ALTER TABLE " + tableName + " ALTER COLUMN ID RESTART WITH 1").executeUpdate();
-        }
-
-        entityManager.createNativeQuery("SET REFERENTIAL_INTEGRITY TRUE").executeUpdate();
-    }
-}
